Screen Capture Performance Improvements
Wrote a .dll to use the Window DXGI Desktop Duplication API to more effeciently grab and scale images to the bits needed for the Visualight visualizer.
This is approximately 30+ times faster than the old brute force method.
Also tweaked a few default parameters for Visualight to be a bit closer to perceived on-screen colors (at least on my LED strips)

Fixed the No GUI option to work properly.  With this option turned off, most controls are still available via MQTT/Home Assistant, but it reduces CPU overhead significantly.

With the GUI turned off and using the new capture method, it's now viable to use in combination with demanding games at relatively high capture rates.