Added Minimum Level and Custom Display Topic
Minimum Level for Visualight visualizer clamps the minimum output amount so it will never go entirely black.
Custom Display Topic - this is an MQTT topic that will be monitored for handling turning on/off screen capture.  If the local monitoring won't work because it goes through a receiver or whatever, this can be used to turn on/off screen capture externally.